Extract Rust ad4m-client lib from cli
