+++ rust-client/schema.gql
