+++ rust-client/src/agent.gql
