+++ rust-client/src/neighbourhoods.gql
