+++ rust-client/src/perspectives.gql
