+++ rust-client/src/types.rs
